# apifox.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "实时协作任务看板 API",
    "description": "Next.js App Router + Prisma 7 + NextAuth v5 实现的项目看板后端接口。所有接口基于 Cookie 会话（NextAuth）鉴权，统一响应格式为 { success, data?, error? }。看板列为固定的四种任务状态（TaskState）：TODO=未开始、UP_NEXT=接下来、IN_PROGRESS=进行中、DONE=已完成。",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发环境"
    }
  ],
...
  "security": [
    { "sessionCookie": [] }
  ],
  "paths": {
    "/api/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "获取当前用户参与的项目列表",
        "description": "返回当前用户作为成员参与的所有项目，含所有者、成员列表和当前用户角色，按加入时间倒序。",
        "responses": {
          "200": {
            "description": "项目列表",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseProjectList" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "创建项目",
        "description": "创建项目并自动将当前用户设为 ADMIN 成员。看板列为固定的四种任务状态，不随项目创建。返回项目详情（含 role 和 activityCount）。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProjectRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseProjectDetail" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token",
        "description": "NextAuth v5 会话 Cookie。生产环境（HTTPS）下名称为 __Secure-authjs.session-token。请先通过 /api/auth 登录后由浏览器自动携带。"
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiErrorResponse" },
            "example": { "success": false, "error": "缺少必要参数" }
          }
        }
      },
      "Unauthorized": {
        "description": "未登录",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiErrorResponse" },
            "example": { "success": false, "error": "未登录" }
          }
        }
      },
...
      "ServerError": {
        "description": "服务器内部错误",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiErrorResponse" },
            "example": { "success": false, "error": "操作失败" }
          }
        }
      }
    },
    "schemas": {
...
      "MemberRole": {
        "type": "string",
        "enum": ["ADMIN", "MEMBER"],
        "description": "项目成员角色"
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string", "nullable": true },
          "email": { "type": "string", "nullable": true },
          "image": { "type": "string", "nullable": true }
        }
      },
      "UserBrief": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string", "nullable": true },
          "image": { "type": "string", "nullable": true }
        }
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "ownerId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ProjectListItem": {
        "allOf": [
          { "$ref": "#/components/schemas/Project" },
          {
            "type": "object",
            "properties": {
              "role": { "$ref": "#/components/schemas/MemberRole" },
              "owner": { "$ref": "#/components/schemas/UserBrief" },
              "members": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Member" }
              }
            }
          }
        ]
      },
      "ProjectDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Project" },
          {
            "type": "object",
            "properties": {
              "role": { "$ref": "#/components/schemas/MemberRole" },
              "activityCount": { "type": "integer", "description": "项目活动记录总数" },
              "lastReadAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true,
                "description": "当前请求者对项目级讨论的已读水位线（值为已读到的最后一条评论的 createdAt，null 表示从未读过）。只属于请求者本人，不随 members 下发"
              },
              "firstUnreadCommentId": {
                "type": "string",
                "nullable": true,
                "description": "打开项目时的未读起点：项目级讨论里第一条「晚于本人水位线、未解决、不是自己发的」评论 id（null 表示没有未读）。前端据此在第一条未读之前画「以下为未读消息」分割线，本次会话内固定不变"
              },
              "owner": { "$ref": "#/components/schemas/UserBrief" },
              "members": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Member" }
              }
            }
          }
        ]
      },
      "Member": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "role": { "$ref": "#/components/schemas/MemberRole" },
          "joinedAt": { "type": "string", "format": "date-time" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
...
      "CreateProjectRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "example": "我的项目", "description": "项目名称，不能为空" },
          "description": { "type": "string", "nullable": true, "example": "项目描述" }
        }
      },
...
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string" }
        }
      },
...
      "ApiResponseProjectDetail": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "$ref": "#/components/schemas/ProjectDetail" }
        }
      },
      "ApiResponseProjectList": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProjectListItem" }
          }
        }
      },
...
    }
  }
}